changed where state lies in app, added friend.js
