+++ friends/src/App.css
@@ -1,22 +1,17 @@
 .App {
   text-align: center;
 }
-
+header{
+  display:flex;
+  justify-content: space-around;
+}
+h2 {
+  margin:0;
+}
 .App-logo {
   animation: App-logo-spin infinite 20s linear;
   height: 40vmin;
   pointer-events: none;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
 }
 
 .App-link {
@@ -34,8 +29,9 @@
 
 .friend-list{
   display:flex;
-  margin:10px 20px;
+  margin:10px auto;
   background:#cdcdcd;
+  width:70%;
  
 }
 .friend-list p{
@@ -52,4 +48,6 @@
 .friend-form input{
   font-size:1.5rem;
 }
-.btn{}
+.btn{
+  padding:5px;
+}
